refactor(exception): 디렉토리 이름 변경
Before : common 디렉토리에 ErrorMessage와 RacingCarException이 존재

After : 예외처리 관련 클래스 이외 다른 클래스들이 없으므로 디렉토리 명을 exception으로 변경
