+++ src/main/java/racingcar/exception/ErrorMessage.java
@@ -1,4 +1,4 @@
-package racingcar.common;
+package racingcar.exception;
 
 public enum ErrorMessage {
